Fix pod install in ci_post_clone.sh

+++ ios/ci_scripts/ci_post_clone.sh
@@ -16,11 +16,15 @@ flutter precache --ios
 # Install Flutter dependencies.
 flutter pub get
 
-# Install CocoaPods using Homebrew.
-HOMEBREW_NO_AUTO_UPDATE=1 # disable homebrew's automatic updates.
-brew install cocoapods
-
 # Install CocoaPods dependencies.
-cd ios && pod install # run `pod install` in the `ios` directory.
+# Check if Podfile exists before running pod install
+if [ -f "ios/Podfile" ]; then
+    # Install CocoaPods using Homebrew.
+    HOMEBREW_NO_AUTO_UPDATE=1 # disable homebrew's automatic updates.
+    brew install cocoapods
+
+    # run `pod install` in the `ios` directory.
+    cd ios && pod install
+fi
 
 exit 0
